Se modifica el retorno de Double a BigDecimal y se modifican los test acorde a este cambio. También se realiza un pequeño cambio en la parte de swagger del readme.md

+++ src/main/java/org/example/prueba/domain/service/impl/CalculadoraServiceImpl.java
@@ -23,7 +23,7 @@ public class CalculadoraServiceImpl implements ICalculadoraService {
      * @return El resultado dependiendo del tipo de operacion introducido
      */
     @Override
-    public double calculo(BigDecimal primerNumero, BigDecimal segundoNumero, String tipoOperacion) {
+    public BigDecimal calculo(BigDecimal primerNumero, BigDecimal segundoNumero, String tipoOperacion) {
 
         if (LOGGER.isDebugEnabled()) {
             LOGGER.debug("Calculando resultado para : {} {} {}", primerNumero, segundoNumero, tipoOperacion);
@@ -34,9 +34,9 @@ public class CalculadoraServiceImpl implements ICalculadoraService {
         if (operador != null) {
             switch (operador) {
                 case SUMA:
-                    return primerNumero.add(segundoNumero).doubleValue();
+                    return primerNumero.add(segundoNumero);
                 case RESTA:
-                    return primerNumero.subtract(segundoNumero).doubleValue();
+                    return primerNumero.subtract(segundoNumero);
                 default:
                     if (LOGGER.isErrorEnabled()) {
                         LOGGER.error("No soporta el operador: {}", operador);
